Add basic framework of files for CI/CD
For continious intergreation and deployment a framework of
Dockerfiles and workflows were needed. We also would require
files for that workflow to run.

I've added three main python files, client, api and server.
Each of these will represent a portion of the project and
are referenced in the new Dockerfile. I've also added the
workflow for GitHub actions as well as a .gitignore.