Add a git commit template file
To give an indication of what each commit message should contain and
also to allow for consistancy and compatibility we should have add a git
commit template message.

I've added a basic git template message in .gitignore that also contains
instructions on how to configure a local repo so that it is pro-loaded
when commiting any changes.